[clangd] Build syntax tree for tweaks
We end up building a new instance on each tweak request, which is not
ideal. We should instead prefer to build a single syntax tree with
ParsedAST and re-use it later.

The problem is that the tree can be mutated in Tweak::apply and newer
requests can fail because of this. To address this limitation, we will
probably need to allow rolling back changes made to syntax trees.