[AST] Add a flag indicating if any subexpression had errors
Summary:
The only subexpression that is considered an error now is TypoExpr, but
we plan to add expressions with errors to improve editor tooling on broken
code. We intend to use the same mechanism to guard against spurious
diagnostics on those as well.

See the follow-up revision for an actual usage of the flag.

Reviewers: rsmith

Subscribers: nridge, merge_guards_bot, aaron.ballman, riccibruno, jfb, cfe-commits

Tags: #clang

Differential Revision: https://reviews.llvm.org/D65591